+++ docs/scripts/sync_notebooks.sh
@@ -7,6 +7,7 @@ mkdir -p docs/source/notebooks/
 mv fdtdx-notebooks-main/*.ipynb docs/source/notebooks/
 mv fdtdx-notebooks-main/quickstart/*.ipynb docs/source/notebooks/
 mv fdtdx-notebooks-main/advanced/*.ipynb docs/source/notebooks/
+mv fdtdx-notebooks-main/examples/*.ipynb docs/source/notebooks/
 
 # Clean up
 rm repo.zip